requirements.txt
django
djangorestframework
pillow
stripe
google-auth